Refactored CLI new commands to put questionnaires in own files
